+++ CHANGELOG.md
@@ -1,5 +1,18 @@
 # Changelog
 
+## Version 2.2.2 (2023-08-15)
+
+### Improvement
+
+- convertTableToExcel: related to hight & width
+- convertTableToExcel get optinal function rowHeightScaleFunction & colWidthScaleFunction
+
+## Version 2.1.2 (2023-08-14)
+
+### Bug Fixes
+
+-  Fixed issue with incorrect CDN address.
+
 ## Version 2.1.0 (2023-08-14)
 
 ### New Features